refactor(chat): merge codex-character-chat into agent-providers
Move the experimental Codex character chat CLI to
agent-providers/character-chat.js and fold its README into the
agent-providers README, since both examples share the same
@aituber-onair/chat/agent + codex-sdk setup. Also consolidate the
former agent-provider-example.js into index.js, removing the
five-line wrapper indirection.
