+++ packages/chat/examples/README.md
@@ -74,14 +74,9 @@ JavaScript runtime examples for Agent SDK providers:
 - Supports `codex-sdk`, `claude-agent-sdk`, and `copilot-sdk`
 - Requires the corresponding SDK package and local SDK authentication
 - Text chat only; browser/GAS/UMD environments are not supported
-
-### [Codex Character Chat](./codex-character-chat/)
-Lightweight experimental CLI for using Codex SDK as an AI character chat engine:
-
-- Uses `@aituber-onair/chat/agent` with the `codex-sdk` provider
-- Keeps a short conversation history in the terminal
-- Supports a custom character name and system prompt
-- Text-only proof of concept for agent-SDK-powered AI characters
+- Includes `character-chat.js`, a lightweight experimental CLI that uses the
+  Codex SDK as an AI character chat engine with a custom character name,
+  system prompt, and short terminal conversation history
 
 ### [AITuber Secretary Agent Example](./character-agent/)
 TypeScript CLI example for an AI character secretary:
